Worked on front end today, it's mostly done.
The UI is looking good so far, I made the home page and a tracking page which is the only page visible for now since the Links aren't in place yet. I also made a progress showing page called "Stats" and It's simple but looks good. I'll wrap this up tomorrow by connecting from end to back end and making sure the link navigation works.