Updates to the tracking page and Timer page
I worked on moving the state up from Timer to Tracking, so that when the timer is done, the tracking page is alerted and can try a post request to upload the data to the database. The route for sessions looks good, tomorrow I'll test this and then work on the stats page.