Tested session route through Postman
I was able to send and store some mock study session data to MongoDB Atlas. I then worked on the front end, incorporation MUI styles and removing custom css. The layout is coming together well, I just need to fix some overflow issues for the timer. Once that's done I can test the timer and see if the front-end is working with the server