Improvments of Home page
I made some changes to the UI of the home page. Now the only things visible are the project name input and the timer. I moved the custom time input to the settings page. The settings page is in charge of setting the time for pomodoro sessions and short and long breaks. Once saved, the home page will have these values passed as props. I'm still working on completing the UI, I used gemini to help me implement MUI styles, now I'm tweaking the layout to my preferences