Finished Home Page Styling
The styling and layout of the home page is mostly finished. I added a 2D Pixel art background to the app as well as an intro scene where the app is hidden behind a 'Begin' button which then fades out to see the background and then shows the app. I'll keep working on the styles for the rest of the pages.